packages/agent-connect/pyproject.toml
[project]
name = "agent-connect"
version = "0.1.0"
description = "Connect AI agents to SaaS — one YAML file per service. OAuth, token vault, tool discovery."
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
keywords = ["agent", "ai", "saas", "oauth", "connector", "mcp", "tools"]
authors = [
    {name = "Eddy Xu"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "cryptography>=41.0",
    "httpx>=0.28",
    "fastapi>=0.129",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_connect"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
